Move tagging helpers into dedicated module_utils.tagging
They're directly ec2 related, ec2 was just a dumping ground.
