+++ plugins/module_utils/ec2.py
@@ -50,6 +50,10 @@ from ansible.module_utils.common.dict_transformations import camel_dict_to_snake
 from ansible.module_utils.common.dict_transformations import snake_dict_to_camel_dict  # pylint: disable=unused-import
 
 from .cloud import CloudRetry
+# Used to live here, moved into ansible_collections.amazon.aws.plugins.module_utils.tagging
+from .tagging import ansible_dict_to_boto3_tag_list
+from .tagging import boto3_tag_list_to_ansible_dict
+from .tagging import compare_aws_tags
 
 BOTO_IMP_ERR = None
 try:
@@ -462,73 +466,6 @@ def ansible_dict_to_boto3_filter_list(filters_dict):
     return filters_list
 
 
-def boto3_tag_list_to_ansible_dict(tags_list, tag_name_key_name=None, tag_value_key_name=None):
-
-    """ Convert a boto3 list of resource tags to a flat dict of key:value pairs
-    Args:
-        tags_list (list): List of dicts representing AWS tags.
-        tag_name_key_name (str): Value to use as the key for all tag keys (useful because boto3 doesn't always use "Key")
-        tag_value_key_name (str): Value to use as the key for all tag values (useful because boto3 doesn't always use "Value")
-    Basic Usage:
-        >>> tags_list = [{'Key': 'MyTagKey', 'Value': 'MyTagValue'}]
-        >>> boto3_tag_list_to_ansible_dict(tags_list)
-        [
-            {
-                'Key': 'MyTagKey',
-                'Value': 'MyTagValue'
-            }
-        ]
-    Returns:
-        Dict: Dict of key:value pairs representing AWS tags
-         {
-            'MyTagKey': 'MyTagValue',
-        }
-    """
-
-    if tag_name_key_name and tag_value_key_name:
-        tag_candidates = {tag_name_key_name: tag_value_key_name}
-    else:
-        tag_candidates = {'key': 'value', 'Key': 'Value'}
-
-    # minio seems to return [{}] as an empty tags_list
-    if not tags_list or not any(tag for tag in tags_list):
-        return {}
-    for k, v in tag_candidates.items():
-        if k in tags_list[0] and v in tags_list[0]:
-            return dict((tag[k], tag[v]) for tag in tags_list)
-    raise ValueError("Couldn't find tag key (candidates %s) in tag list %s" % (str(tag_candidates), str(tags_list)))
-
-
-def ansible_dict_to_boto3_tag_list(tags_dict, tag_name_key_name='Key', tag_value_key_name='Value'):
-
-    """ Convert a flat dict of key:value pairs representing AWS resource tags to a boto3 list of dicts
-    Args:
-        tags_dict (dict): Dict representing AWS resource tags.
-        tag_name_key_name (str): Value to use as the key for all tag keys (useful because boto3 doesn't always use "Key")
-        tag_value_key_name (str): Value to use as the key for all tag values (useful because boto3 doesn't always use "Value")
-    Basic Usage:
-        >>> tags_dict = {'MyTagKey': 'MyTagValue'}
-        >>> ansible_dict_to_boto3_tag_list(tags_dict)
-        {
-            'MyTagKey': 'MyTagValue'
-        }
-    Returns:
-        List: List of dicts containing tag keys and values
-        [
-            {
-                'Key': 'MyTagKey',
-                'Value': 'MyTagValue'
-            }
-        ]
-    """
-
-    tags_list = []
-    for k, v in tags_dict.items():
-        tags_list.append({tag_name_key_name: k, tag_value_key_name: to_native(v)})
-
-    return tags_list
-
-
 def get_ec2_security_group_ids_from_names(sec_group_list, ec2_connection, vpc_id=None, boto3=True):
 
     """ Return list of security group IDs from security group names. Note that security group names are not unique
@@ -778,33 +715,6 @@ def map_complex_type(complex_type, type_map):
     elif type_map:
         return globals()['__builtins__'][type_map](complex_type)
     return new_type
-
-
-def compare_aws_tags(current_tags_dict, new_tags_dict, purge_tags=True):
-    """
-    Compare two dicts of AWS tags. Dicts are expected to of been created using 'boto3_tag_list_to_ansible_dict' helper function.
-    Two dicts are returned - the first is tags to be set, the second is any tags to remove. Since the AWS APIs differ
-    these may not be able to be used out of the box.
-
-    :param current_tags_dict:
-    :param new_tags_dict:
-    :param purge_tags:
-    :return: tag_key_value_pairs_to_set: a dict of key value pairs that need to be set in AWS. If all tags are identical this dict will be empty
-    :return: tag_keys_to_unset: a list of key names (type str) that need to be unset in AWS. If no tags need to be unset this list will be empty
-    """
-
-    tag_key_value_pairs_to_set = {}
-    tag_keys_to_unset = []
-
-    for key in current_tags_dict.keys():
-        if key not in new_tags_dict and purge_tags:
-            tag_keys_to_unset.append(key)
-
-    for key in set(new_tags_dict.keys()) - set(tag_keys_to_unset):
-        if to_text(new_tags_dict[key]) != current_tags_dict.get(key):
-            tag_key_value_pairs_to_set[key] = new_tags_dict[key]
-
-    return tag_key_value_pairs_to_set, tag_keys_to_unset
 
 
 @AWSRetry.jittered_backoff()
